requirements.txt
boto3==1.24.35
covalent-aws-plugins==0.5.0rc0
